+++ Sibusten.Philomena.Client/IPhilomenaImage.cs
@@ -15,8 +15,8 @@ namespace Sibusten.Philomena.Client
         ImageModel Model { get; }
 
         int Id { get; }
-        string Name { get; }
-        string OriginalName { get; }
+        string? Name { get; }
+        string? OriginalName { get; }
 
         /// <summary>
         /// Downloads the image